requirements.txt
# Core Data Processing
pandas==2.1.4
numpy==1.24.3
openpyxl==3.1.2

# Machine Learning & Prediction
scikit-learn==1.3.2
prophet==1.1.5
statsmodels==0.14.1

# RAG Components
chromadb==0.4.22
sentence-transformers==2.3.1
langchain==0.1.0
langchain-community==0.0.13

# LLM Integration (using Gemini like your previous work)
google-generativeai==0.3.2

# OpenTelemetry for Tracing
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-exporter-jaeger==1.22.0
opentelemetry-instrumentation==0.43b0

# Visualization & Monitoring
matplotlib==3.8.2
plotly==5.18.0

# API & Web (optional for deployment)
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3

# Configuration
pyyaml==6.0.1
python-dotenv==1.0.0

# Testing
pytest==7.4.4
